Minor bug fix

+++ scripts/run_analytics.sh
@@ -121,6 +121,8 @@ if [ -n "$1" ]; then
         fi
     fi
     if [ "$1" = "radixgraph" ]; then
-        rm settings.txt
+        if [ -e "settings.txt" ]; then
+            rm settings.txt
+        fi
     fi
 fi
